\name{Rglimclim-Fortran}
\alias{Rglimclim-Fortran}
\alias{closefiles}
\alias{dailystats}
\alias{glmfit}
\alias{readdata2}
\alias{rglcfileconnect}
\alias{scratchsync}
\alias{scratchupdate}
\alias{scropn}
\alias{simulate}
\alias{writedata}
\alias{writeheader}
\alias{writelabels}
\alias{zbqlini}
\title{Rglimclim Fortran routines}
\description{These objects are the routines in the underlying \code{Fortran} code. They are not intended to be called by the user.}
\keyword{Fortran}
