\name{Rglimclim-internal}
\alias{Rglimclim-internal}
\alias{file.OK}
\alias{reformat.scodes}
\alias{err.msg}
\alias{expand.pwtidx}
\alias{orth.set}
\alias{lagweights.def}
\alias{model.num}
\alias{make.labels}
\alias{char.trunc}
\alias{title.print}
\alias{OpenFile}
\alias{OpenTempFile}
\alias{CloseFiles}
\alias{CheckFiles}
\alias{GLCfit.internal}
\alias{CheckJointModel}
\alias{JointFit.internal}
\alias{get.nesting}
\alias{set.maxlag}
\alias{dep.calc}
\alias{check.siteinfo}
\alias{make.regnames}
\alias{make.simfilename}
\alias{rho.check}
\alias{circles}
\alias{write.CharVecAsTable}
\alias{quantileplot}
\title{Rglimclim internal functions}
\description{These functions are provided mainly for internal housekeeping purposes and for smooth interfacing with the underlying \code{Fortran} code. They are not intended to be called by the user.}
\keyword{internal}
